requirements.txt
Flask-RESTful==0.3.9
Flask==2.1.0
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.34
python-dotenv==0.20.0
